rename CommandVars to ModConfig (IMHO more expressive name)
